Volta

+++ src/ClasseCaminhoes.java
@@ -3,7 +3,7 @@ public class ClasseCaminhoes extends ClasseAutomotores{
 	float cargaSuportadaAt;
 	
 	//super
-	public ClasseCaminhoes(String corPar, String marcaPar, String modeloPar, String tipoPar, int cargaSuportadaPar) {
+	public ClasseCaminhoes(String corPar, String marcaPar, String modeloPar, String tipoPar, float cargaSuportadaPar) {
 		super(corPar, marcaPar, modeloPar, tipoPar);
